Add new items, extend targeting system
Currently the targeting rendering system is pretty hodgepodge, but at
least its pretty!

It's based off of dcss's targeting ui, except that it requires the
cursor to be on the target: it doesn't let you draw past the enemy to
shoot it. That might be better, but would need a little bit of
refactoring.